StudySafeCore/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.4
django-filter==21.1
django-stubs==1.10.1
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-stubs==1.5.0
idna==3.3
importlib-metadata==4.11.3
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
mypy==0.942
mypy-extensions==0.4.3
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
types-Markdown==3.3.13
types-pytz==2021.3.6
types-PyYAML==6.0.6
types-requests==2.27.19
types-urllib3==1.26.13
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
